Power Query Excercise.....More to come
I am extending my experience of Excel used in many of my contracting roles over twenty years.  I am very interested in Data Analytics.  The world of Big Data itself is not perfect.  It needs organizing and transforming before it can be analyzed and made useful for consumption.  Watch this space!! I hope to be displaying more of my skills in this particular area.